+++ src/static_memory_manager/staticMemoryManager.class.hpp
@@ -35,7 +35,7 @@ public:
 
 	static uint64_t									create_slot_child(E_ASSET_TYPE type);
 
-	static uint64_t									create_asset(uint32_t cluster_id, uint32_t size);
+	static uint32_t									create_asset(uint32_t cluster_id, uint32_t size);
 
 	static staticMemoryManager::e_asset_state		get_asset_state(uint64_t ref);
